+++ dialogExecution/newVirtualMachine.py
@@ -292,10 +292,10 @@ class NewVirtualMachineDialog(QDialog, Ui_Dialog):
 
         add_args = self.lineEdit_2.text()
 
-        if self.checkBox_3.isChecked() or self.checkBox.isChecked() or self.comboBox_13.currentText("USB Mouse"):
+        if self.checkBox_3.isChecked() or self.checkBox.isChecked() or self.comboBox_13.currentText() == "USB Mouse":
             usb_support = 1
 
-        elif self.comboBox_13.currentText("USB Tablet Device") or self.comboBox_16.currentText("USB Keyboard"):
+        elif self.comboBox_13.currentText() == "USB Tablet Device" or self.comboBox_16.currentText() == "USB Keyboard":
             usb_support = 1
 
         else: