+++ main.py
@@ -21,7 +21,7 @@ class Window(QMainWindow, Ui_MainWindow):
         self.connectSignalsSlots()
         self.timer = QTimer()
         self.timer.timeout.connect(self.updateVmList)
-        self.label_8.setText("EmuGUI v0.4.1")
+        self.label_8.setText("EmuGUI v0.4.2")
         self.setWindowTitle("EmuGUI")
 
         if platform.system() == "Windows":